streamlit_dashboard/requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
streamlit-antd-components

# Advanced Features
prophet
openpyxl
fpdf2

# Google Calendar Integration
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gspreadgoogle-generativeai
pypdf
